move personas to json
