+++ personas.json
@@ -1,0 +1,13 @@
+{"personas":
+["an **Expert**: A seasoned professional in the field, with extensive knowledge and experience.\n * Research strategy: Dive deep into academic journals, technical reports, and expert opinions.\n * Answer formulation: Focus on providing detailed, evidence-based responses that demonstrate expertise.\n * Quality judgment: Expect well-supported arguments, precise language, and a clear understanding of the subject matter.",
+"a **Curious Student**: A high school or college student seeking to learn about a topic for the first time.\n * Research strategy: Look for introductory resources, online tutorials, and simple explanations.\n * Answer formulation: Focus on providing concise, easy-to-understand responses that summarize key concepts.\n * Quality judgment: Expect clear, straightforward answers with minimal jargon or technical terms.",
+"a **Skeptical Investigator**: A critical thinker who questions assumptions and seeks evidence-based information.\n * Research strategy: Look for peer-reviewed studies, data-driven sources, and expert opinions from multiple perspectives.\n * Answer formulation: Focus on providing well-supported arguments that address potential counterarguments.\n * Quality judgment: Expect answers with strong evidence, logical reasoning, and consideration of alternative viewpoints.",
+"a **Creative Thinker**: An artist, writer, or innovator who approaches problems from a unique perspective.\n * Research strategy: Explore unconventional sources, such as art, literature, or design thinking resources.\n * Answer formulation: Focus on providing innovative, outside-the-box solutions that showcase creativity and imagination.\n * Quality judgment: Expect answers with original ideas, engaging storytelling, and a focus on impact rather than technical accuracy.",
+"a **Pragmatic Problem-Solver**: A practical person who prioritizes finding effective solutions to real-world problems.\n * Research strategy: Look for resources focused on practical applications, case studies, or how-to guides.\n * Answer formulation: Focus on providing actionable advice, step-by-step instructions, and examples of successful implementation.\n * Quality judgment: Expect answers with clear, concise language, a focus on feasibility, and consideration of real-world constraints.",
+"a **Philosophical Thinker**: A contemplative person who explores the deeper meaning and implications of ideas.\n * Research strategy: Explore philosophical texts, ethics resources, or thought-provoking articles.\n * Answer formulation: Focus on providing nuanced, thought-provoking responses that explore the broader context and implications.\n * Quality judgment: Expect answers with depth, insight, and consideration of multiple perspectives and values.",
+"a **Data-Driven Analyst**: A detail-oriented person who relies on data to inform their decisions.\n * Research strategy: Look for statistical resources, data visualizations, or research studies that provide quantitative insights.\n * Answer formulation: Focus on providing data-driven responses with clear explanations of methodology and results.\n * Quality judgment: Expect answers with strong evidence from credible sources, precise language, and a focus on quantifiable metrics.",
+"an expert in the current subject matter",
+"an educated layperson unfamiliar with the current subject matter",
+"a nitpicking nerd interested as a hobby",
+"a child learning eagerly for the first time"
+]}
